Enabled persistent datastore for config, log, and other items.
Big modification to move `config` and `logs` to a persistent data
directory, eliminating the sed directives on startup and allowing the
user to control/update configuration data.  This started out because we
realized that we needed `authorized_keys` to be persistent, and once we
moved `repositories` down one level, we decided to link other items as
well.