final commit for new build env
Fixed a typo in nginx config.  Pulled the symlinking data out and made
it a configurable file (`links`) in the persistent data directory.  This
will allow us to move other data out to the persistent store at a future
time without having to rebuild the container.  Also fixed a typo in the
start config, which unfortunately means I'll have to rebuild the
container.  Suckage.